requirements.txt
Django==1.8
GDAL==1.11.1
PyYAML==3.11
SPARQLWrapper==1.6.4
argparse==1.3.0
boto==2.34.0
dateutils==0.6.6
djangorestframework==2.3.14
djangorestframework-gis==0.3
futures==2.2.0
html5lib==0.999
isodate==0.5.1
pbr==0.10.0
prettytable==0.7.2
psycopg2==2.5.3
py==1.4.26
pyparsing==2.0.3
pytest==2.7.0
pytest-django==2.8.0
python-dateutil==2.2
pytz==2015.2
rdflib==4.2.0
requests==2.6.2
six==1.9.0
stevedore==1.1.0
tinys3==0.1.11
titlecase==0.7.1
uWSGI==2.0.6
virtualenv==1.11.6
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
wsgiref==0.1.2
